Fix older posts not loading
When going to a post detail page that wasn't loaded as of normal
page load, the page would be blank. Need to fetch the page from
the backend if it isn't in the cache.

Also fixed an error in dtsince where 'X s' for seconds would show up
as 'X true'.